+++ api/apiFunctions.ts
@@ -1,17 +1,40 @@
-import { account } from "./config"
+import { ID } from 'appwrite';
+import { account } from './config';
 
 export async function loginAccount(user: { email: string; password: string }) {
   try {
-      return await account.createEmailPasswordSession(user.email, user.password);
+    return await account.createEmailPasswordSession(user.email, user.password);
   } catch (error) {
-      console.error(error);
+    console.error(error);
   }
 }
 
 export async function logoutAccount() {
   try {
-      return await account.deleteSession('current');
+    return await account.deleteSession('current');
   } catch (error) {
-      console.error(error);
+    console.error(error);
+  }
+}
+
+export async function POST(request: Request): Promise<Response> {
+  const { email, password } = await request.json();
+
+  try {
+    const response = await account.create(ID.unique(), email, password);
+
+    return new Response(JSON.stringify(response), {
+      status: 200,
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
+  } catch (error: any) {
+    return new Response(JSON.stringify({ error: error.message }), {
+      status: 500,
+      headers: {
+        'Content-Type': 'application/json',
+      },
+    });
   }
 }
